+++ lua/primebrook/oil.lua
@@ -1,1 +1,7 @@
-require("oil").setup()
+local M = {}
+
+function M.setup()
+	require("oil").setup()
+end
+
+return M
